fix: 622xxx UnionPay 카드 인식 버그 수정 및 네트워크 감지 모듈 응집
