+++ my-app/src/utils/validators.ts
@@ -7,17 +7,11 @@ export const validateNumeric = (value: string): string => {
   return "";
 };
 
-export const validateCardNumbers = (
-  values: string[],
-  isSupportedNetwork: boolean,
-): { index: number; message: string } => {
+export const validateCardNumbers = (values: string[]): { index: number; message: string } => {
   for (const [i, value] of values.entries()) {
     if (!value) continue;
     if (!NUMERIC_REGEX.test(value)) {
       return { index: i, message: "숫자만 입력 가능합니다" };
-    }
-    if (i === 0 && !isSupportedNetwork) {
-      return { index: i, message: "이 카드 브랜드는 지원하지 않습니다." };
     }
   }
   return { index: -1, message: "" };
